Add Neovim/Vim related info to Development.md (#124)
This commit adds some information about setting up Neovim or Vim as a
development tool for the Swift driver. It details the process for
setting up `nvim` or `vim` to work on Swift codebases with proper
indenting and syntax support, as well as how to set up linting inside of
`nvim` or `vim` via `ale` and `swiftlint`.