SWIFT-201 Refactor "if failure" conditions with guard statements (#125)
This commit refactors multiple instances where we use an `if` statement to effectively guard against some exceptional case, usually a failure. In these kinds of situations, it is better to use the `guard` keyword in Swift to do the same job, as it is generally cleaner and saves on indentation, etc. We therefore refactor these instances of `if` to use `guard`.