Add an implementation for bsonEqual()
This commit adds an implementation for bsonEqual(), a helper function
for comparing two arbitrary BSONValues for equality. This function tests
for hard equality, so the BSONValues must be of the same type.

This function's implementation also 'ignores' Symbol, DBPointer and
Array.

The first two are done for the same reason, being that they are
deprecated symbols and users should not be making instances of them
anyways.

The third type, Array, is disabled because as far as I can tell, there
is no way to be able to take any arbitrary array with arbitrary type and
confirm that the given type conforms to Equatable and then act
accordingly depending on whether or not it does.