Throw error instead of crash for invalid arrays
Before, we were tossing preconditionFailure when we saw an invalid
array, aka, not a [BSONValue?]. This is pretty dumb because a user
should not have their program crash just because their argument is
invalid. The other precondition failures are okay, since they should
basically never happen unless the user is doing something really dumb.