SWIFT-251, SWIFT-247: Create an issue template on the GitHub repo, Create an example project for bug submission (#142)
This commit adds an issue template, as well as a PR template and a CONTRIBUTING.md file. This should help with people who contribute to the repository.

This commit also does the work of SWIFT-247 and adds a sample project to which users can add code in the main Swift file to reproduce bugs.