Defer creation of collection to operation execution
A few things going on here. First, a new type `MongoNamespace` was
introduced to make it easy to track the current namespace without
needing to necessarily pass collections or databases around.
Secondly, execution is now passed a pointer to the `mongoc_client_t`
so that we can create the `mongoc_collection_t` or `mongoc_database_t`
at the moment of execution, this allows us to avoid most of the
pitfalls of lack of thread safety in libmongoc.