Introduce a client pool
This changes introduces a client pool into the `MongoClient` ctor.
Of note is that the pool is currently only used for two operations
which have been converted to `Operation` subclasses (insertOne,
updateOne), and it maintains behavior for all other operations by
popping a client at init, and pushing it back before close.