SWIFT-399: enable evg builds on OSX hosts
This is dependent on a branch of `swiftenv` on my local github
account to work, that will hopefully be resolved shortly. Other
items of note here are that we only build on the macos-1014 hosts
due to XCode version requirements, and some tricky linker lines
to accommodate rpath linking to libmongoc.